marine-plastic-detection/requirements.txt
Flask
torch
torchvision
numpy
opencv-python
pillow
requests
flask-cors
pyyaml
matplotlib
scikit-learn
tensorflow
tensorflow-hub
imageio
imageio[ffmpeg]